+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Dropdown menu fully closes when clicking on avatar.
 * Added setting to settings bundle.
 * Adjusted design of app to fit on 4.7inch displays (for development)
+* Recording playback stops when you exit from the playback view controller.
 ## v0.2.5 (PRERELEASE)
 * Account settings editing now works (actually contacts the api)
 * Fixed bug where timer would stop forever and load any user info until app restart